requirements.txt
requests
schedule
kafka-python
pymongo
